fix: Don't allow multi-line strings to indent by more than 2 spaces
The `tomelr-indent-multi-line-strings` variable is set to `t' mainly
by ox-hugo. So these multi-line strings will most likely be parsed by
Hugo as Markdown. If the indentation of lines in these strings is 4
spaces or more, they get parsed as code blocks! So limit the
indentation of these strings to just 2 spaces.